Refactoring to improve code consistency
- Removing any patterns introduced that weren't already established in the repo
- Removing hooks with database reads in favor of withObservables and withDatabase pattern
- Simplifying _layout.tsx by removing inset modification, now handled in the Container HoC component
- Moved some repeated hook logic into its own utils file
