+++ app/queries/servers/properties.ts
@@ -1,12 +1,15 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {of as of$} from 'rxjs';
-import {switchMap} from 'rxjs/operators';
+import {combineLatest, of as of$} from 'rxjs';
+import {distinctUntilChanged, map, switchMap} from 'rxjs/operators';
 
+import {CLASSIFICATIONS_GROUP_NAME, CLASSIFICATIONS_SYSTEM_VALUE_TARGET_ID} from '@constants/classification';
 import {MM_TABLES, SYSTEM_IDENTIFIERS} from '@constants/database';
 import {querySystemValue} from '@queries/servers/system';
+import {deriveChannelClassificationBanner, deriveClassificationBannerState} from '@utils/classification';
 
+import type ServerDataOperator from '@database/operator/server_data_operator';
 import type {Database} from '@nozbe/watermelondb';
 import type SystemModel from '@typings/database/models/servers/system';
 
@@ -66,3 +69,59 @@ export const observePropertyGroupNames = (database: Database) => {
         )),
     );
 };
+
+// --- Classification banner observables ---
+
+export const observeClassificationBannerState = (database: Database) => {
+    return combineLatest([
+        observePropertyFields(database),
+        observePropertyValues(database),
+        observePropertyGroupNames(database),
+    ]).pipe(
+        map(([fields, values, groupNames]) => deriveClassificationBannerState(fields, values, groupNames)),
+        distinctUntilChanged((a, b) => a.visible === b.visible && a.levelName === b.levelName && a.color === b.color),
+    );
+};
+
+export const observeChannelClassificationBanner = (database: Database, channelId: string, nativeBannerText?: string) => {
+    return combineLatest([
+        observePropertyFields(database),
+        observePropertyValues(database),
+        observePropertyGroupNames(database),
+    ]).pipe(
+        map(([fields, values, groupNames]) => deriveChannelClassificationBanner(fields, values, groupNames, channelId, nativeBannerText)),
+        distinctUntilChanged((a, b) => a.hasClassification === b.hasClassification &&
+            a.classificationBanner?.text === b.classificationBanner?.text &&
+            a.classificationBanner?.background_color === b.classificationBanner?.background_color),
+    );
+};
+
+// --- Mutations ---
+
+export async function clearClassificationData(operator: ServerDataOperator) {
+    const {database} = operator;
+
+    const [existingFields, existingValues, existingGroupNames] = await Promise.all([
+        getPersistedPropertyFields(database),
+        getPersistedPropertyValues(database),
+        getPersistedPropertyGroupNames(database),
+    ]);
+
+    const groupId = existingGroupNames[CLASSIFICATIONS_GROUP_NAME];
+    if (!groupId) {
+        return;
+    }
+
+    const {[groupId]: _, ...remainingFields} = existingFields;
+    const {[CLASSIFICATIONS_SYSTEM_VALUE_TARGET_ID]: __, ...remainingValues} = existingValues;
+    const {[CLASSIFICATIONS_GROUP_NAME]: ___, ...remainingGroupNames} = existingGroupNames;
+
+    await operator.handleSystem({
+        systems: [
+            {id: SYSTEM_IDENTIFIERS.PROPERTY_FIELDS, value: remainingFields},
+            {id: SYSTEM_IDENTIFIERS.PROPERTY_VALUES, value: remainingValues},
+            {id: SYSTEM_IDENTIFIERS.PROPERTY_GROUP_NAMES, value: remainingGroupNames},
+        ],
+        prepareRecordsOnly: false,
+    });
+}
